Move isConfigAtHomeRoot to its own file
It needs to be used in both worker and non-worker, so it’s easiest
to just put it in its own file for now until we re-structure.
