+++ lib/helpers.js
@@ -159,20 +159,9 @@ let processESLintMessages = exports.processESLintMessages = (() => {
   };
 })();
 
-/**
- * Check if a config is directly inside a user's home directory.
- * Such config files are used by ESLint as a fallback, only for situations
- * when there is no other config file between a file being linted and root.
- *
- * @param  {string}  configPath - The path of the config file being checked
- * @return {Boolean}            [description]
- */
-
-
 exports.spawnWorker = spawnWorker;
 exports.showError = showError;
 exports.idsToIgnoredRules = idsToIgnoredRules;
-exports.isConfigAtHomeRoot = isConfigAtHomeRoot;
 
 var _child_process = require('child_process');
 
@@ -191,10 +180,6 @@ var _eslintRuleDocumentation = require('eslint-rule-documentation');
 var _eslintRuleDocumentation2 = _interopRequireDefault(_eslintRuleDocumentation);
 
 var _atomLinter = require('atom-linter');
-
-var _userHome = require('user-home');
-
-var _userHome2 = _interopRequireDefault(_userHome);
 
 var _atom = require('atom');
 
@@ -302,6 +287,4 @@ const generateInvalidTrace = (() => {
   return function generateInvalidTrace(_x5, _x6, _x7, _x8, _x9, _x10, _x11, _x12, _x13, _x14) {
     return _ref3.apply(this, arguments);
   };
-})();function isConfigAtHomeRoot(configPath) {
-  return (0, _path.dirname)(configPath) === _userHome2.default;
-}
+})();
