+++ lib/main.js
@@ -11,6 +11,8 @@ var _helpers = require('./helpers');
 
 var _workerHelpers = require('./worker-helpers');
 
+var _isConfigAtHomeRoot = require('./is-config-at-home-root');
+
 function _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }
 
 function _asyncToGenerator(fn) { return function () { var gen = fn.apply(this, arguments); return new Promise(function (resolve, reject) { function step(key, arg) { try { var info = gen[key](arg); var value = info.value; } catch (error) { reject(error); return; } if (info.done) { resolve(value); } else { return Promise.resolve(value).then(function (value) { step("next", value); }, function (err) { step("throw", err); }); } } return step("next"); }); }; }
@@ -58,7 +60,7 @@ module.exports = {
           // Do not try to fix if linting should be disabled
           const fileDir = _path2.default.dirname(filePath);
           const configPath = (0, _workerHelpers.getConfigPath)(fileDir);
-          const noProjectConfig = configPath === null || (0, _helpers.isConfigAtHomeRoot)(configPath);
+          const noProjectConfig = configPath === null || (0, _isConfigAtHomeRoot.isConfigAtHomeRoot)(configPath);
           if (noProjectConfig && disableWhenNoEslintConfig) return;
 
           this.worker.request('job', {
