+++ lib/worker.js
@@ -15,7 +15,7 @@ var _workerHelpers = require('./worker-helpers');
 
 var Helpers = _interopRequireWildcard(_workerHelpers);
 
-var _helpers = require('./helpers');
+var _isConfigAtHomeRoot = require('./is-config-at-home-root');
 
 function _interopRequireWildcard(obj) { if (obj && obj.__esModule) { return obj; } else { var newObj = {}; if (obj != null) { for (var key in obj) { if (Object.prototype.hasOwnProperty.call(obj, key)) newObj[key] = obj[key]; } } newObj.default = obj; return newObj; } }
 
@@ -34,7 +34,7 @@ const ignoredMessages = [
 'File ignored by default.  Use a negated ignore pattern (like "--ignore-pattern \'!<relative' + '/path/to/filename>\'") to override.', 'File ignored by default. Use "--ignore-pattern \'!node_modules/*\'" to override.', 'File ignored by default. Use "--ignore-pattern \'!bower_components/*\'" to override.'];
 
 function lintJob(argv, contents, eslint, configPath, config) {
-  const noProjectConfig = configPath === null || (0, _helpers.isConfigAtHomeRoot)(configPath);
+  const noProjectConfig = configPath === null || (0, _isConfigAtHomeRoot.isConfigAtHomeRoot)(configPath);
   if (noProjectConfig && config.disableWhenNoEslintConfig) {
     return [];
   }
