+++ src/worker.js
@@ -6,7 +6,7 @@ import Path from 'path'
 import { create } from 'process-communication'
 import { FindCache, findCached } from 'atom-linter'
 import * as Helpers from './worker-helpers'
-import { isConfigAtHomeRoot } from './helpers'
+import { isConfigAtHomeRoot } from './is-config-at-home-root'
 
 process.title = 'linter-eslint helper'
